requirements.txt
requests>=2.32.3
google-auth-oauthlib
google-cloud-bigquery